Peer reviewed the changes made by Shubham
Shubham, you mentioned that procedure headings are no longer used. We should discuss this point with Shubhada and team because the style guide suggests using them. If we are deviating from the guideline, the team should have a universal understanding and it should be documented somewhere for reference. I have not added comments for points that we discussed in person.

Thank you.
Neeraja